pyproject.toml
[project]
name = "reservoir-computing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jupyter",
    "yfinance",
    "pandas",
    "numpy",
    "scikit-learn",
    "joblib",
    "matplotlib",
    "beautifulsoup4",
    "lxml",
    "python-dotenv",
]
